feast-example/feature_repo/requirements.txt
snowflake-connector-python==2.7.6
pyarrow==6.0.1
feast-azure-provider==0.3.0
azure-keyvault-secrets
azure-identity